Search page design and logi
Search page pretty much complete with the design and selection logic. Just need to figure out how to send api data results to the Results component to display the shows. Also need to connect user login db details to Search page so that the hello message is personalised to the username.